refactored navbar and added shopping cart in navbar
