+++ frontend/src/components/custom/BuyerNavbar.tsx
@@ -1,64 +1,151 @@
 "use client";
 
-import Navbar from "@/components/custom/Navbar";
+import { Button } from "@/components/ui/button";
 import {
-	NavigationMenu,
-	NavigationMenuItem,
-	NavigationMenuLink,
-	NavigationMenuList,
-	navigationMenuTriggerStyle,
+  Sheet,
+  SheetContent,
+  SheetHeader,
+  SheetTitle,
+  SheetTrigger,
+} from "@/components/ui/sheet";
+import {
+  NavigationMenu,
+  NavigationMenuContent,
+  NavigationMenuItem,
+  NavigationMenuLink,
+  NavigationMenuList,
+  NavigationMenuTrigger,
+  navigationMenuTriggerStyle,
 } from "@/components/ui/navigation-menu";
 import { SheetClose } from "@/components/ui/sheet";
+import { useCartStore } from "@/lib/store/useCartStore";
 import Link from "next/link";
+import { usePathname } from "next/navigation";
+import { useStore } from "zustand";
+import { ShoppingCart } from "./ShoppingCart";
+import { Menu, ShoppingCart as ShoppingCartIcon } from "lucide-react";
+import DesktopUserMenu from "./DesktopUserMenu";
+import MobileUserMenu from "./MobileUserMenu";
 
 export default function BuyerNavbar() {
-	const buyerDesktopButtons = () => {
-		return (
-			<NavigationMenu>
-				<NavigationMenuList>
-					<NavigationMenuItem>
-						<Link href="/buyer/browse" legacyBehavior passHref>
-							<NavigationMenuLink className={navigationMenuTriggerStyle()}>
-								Browse
-							</NavigationMenuLink>
-						</Link>
-					</NavigationMenuItem>
-					<NavigationMenuItem>
-						<Link href="/buyer" legacyBehavior passHref>
-							<NavigationMenuLink className={navigationMenuTriggerStyle()}>
-								Orders
-							</NavigationMenuLink>
-						</Link>
-					</NavigationMenuItem>
-				</NavigationMenuList>
-			</NavigationMenu>
-		);
-	};
-	const buyerMobileButtons = () => {
-		return (
-			<>
-				<SheetClose asChild>
-					<Link
-						href="/buyer"
-						className="flex items-center py-2 text-xl font-semibold">
-						Browse
-					</Link>
-				</SheetClose>
-				<SheetClose asChild>
-					<Link
-						href="/buyer"
-						className="flex items-center py-2 text-xl font-semibold">
-						Orders
-					</Link>
-				</SheetClose>
-			</>
-		);
-	};
-	return (
-		<Navbar
-			userRole="buyer"
-			desktopButtons={buyerDesktopButtons()}
-			mobileButtons={buyerMobileButtons()}
-		/>
-	);
+  const pathname = usePathname();
+
+  const getFundraiserId = () => {
+    // Check if we're on a buyer fundraiser page
+    if (pathname.includes("/buyer/fundraiser/")) {
+      // Split the path by '/' and get the last segment
+      const segments = pathname.split("/");
+      return segments[segments.length - 1];
+    }
+    return undefined;
+  };
+
+  const fundraiserId = getFundraiserId();
+
+  // fixes nextjs hydration issue: https://github.com/pmndrs/zustand/issues/938#issuecomment-1481801942
+  const cart =
+    useStore(useCartStore, (state) => state.carts[fundraiserId ?? ""]) || [];
+
+  return (
+    <header className="sticky top-0 z-50 w-full border-b bg-background">
+      <div className="flex h-16 items-center justify-between px-4 md:px-8 lg:px-12">
+        {/* Logo */}
+        <div className="flex items-center">
+          <Link href="/" className="text-2xl font-bold">
+            CURaise
+          </Link>
+        </div>
+
+        {/* Desktop Navigation - With even spacing */}
+        <div className="hidden md:flex md:flex-1 justify-end">
+          <NavigationMenu>
+            <NavigationMenuList>
+              <NavigationMenuItem>
+                <Link href="/buyer/browse" legacyBehavior passHref>
+                  <NavigationMenuLink className={navigationMenuTriggerStyle()}>
+                    Browse
+                  </NavigationMenuLink>
+                </Link>
+              </NavigationMenuItem>
+              <NavigationMenuItem>
+                <Link href="/buyer" legacyBehavior passHref>
+                  <NavigationMenuLink className={navigationMenuTriggerStyle()}>
+                    Orders
+                  </NavigationMenuLink>
+                </Link>
+              </NavigationMenuItem>
+              {fundraiserId && (
+                <NavigationMenuItem>
+                  <NavigationMenuTrigger>Cart</NavigationMenuTrigger>
+                  <NavigationMenuContent>
+                    <div className="w-[300px] p-4">
+                      <ShoppingCart cart={cart} fundraiserId={fundraiserId} />
+                    </div>
+                  </NavigationMenuContent>
+                </NavigationMenuItem>
+              )}
+            </NavigationMenuList>
+          </NavigationMenu>
+        </div>
+
+        {/* User Menu (Desktop) */}
+        <div className="hidden md:flex md:items-center">
+          <DesktopUserMenu userRole="buyer" />
+        </div>
+
+        {/* Mobile Menu */}
+        <div className="md:hidden">
+          <Sheet>
+            <SheetTrigger asChild>
+              <Button variant="ghost" size="icon" className="md:hidden">
+                <ShoppingCartIcon className="h-5 w-5" />
+                <span className="sr-only">Cart</span>
+              </Button>
+            </SheetTrigger>
+            <SheetContent side="top">
+              <SheetHeader className="mb-4">
+                <SheetTitle className="text-2xl">Cart</SheetTitle>
+              </SheetHeader>
+              <ShoppingCart cart={cart} fundraiserId={fundraiserId} />
+            </SheetContent>
+          </Sheet>
+
+          <Sheet>
+            <SheetTrigger asChild>
+              <Button variant="ghost" size="icon" className="md:hidden">
+                <Menu className="h-5 w-5" />
+                <span className="sr-only">Toggle Menu</span>
+              </Button>
+            </SheetTrigger>
+            <SheetContent side="top">
+              <SheetHeader className="mb-4">
+                <SheetTitle className="text-2xl">CURaise</SheetTitle>
+              </SheetHeader>
+              <div className="grid gap-4 py-4">
+                <SheetClose asChild>
+                  <Link
+                    href="/buyer"
+                    className="flex items-center py-2 text-xl font-semibold"
+                  >
+                    Browse
+                  </Link>
+                </SheetClose>
+
+                <SheetClose asChild>
+                  <Link
+                    href="/buyer"
+                    className="flex items-center py-2 text-xl font-semibold"
+                  >
+                    Orders
+                  </Link>
+                </SheetClose>
+
+                <MobileUserMenu userRole="buyer" />
+              </div>
+            </SheetContent>
+          </Sheet>
+        </div>
+      </div>
+    </header>
+  );
 }
